shell.c: use dev_info() instead of "raw" printk()
It is better to use dev_*() when you have a real device pointer, which
you do here, to better describe the message in the kernel logs.  Convert
the use of printk() to dev_info() to make it easier to see the kernel
messages here.  Bonus is that now you can easily filter for the driver
messages using journalctl and other standard tools.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>